producers/requirements.txt
websocket-client==1.5.1
confluent-kafka==1.9.2
requests==2.31.0